[tfman] move structure browser along with page #2457
