+++ src/man/structure.h
@@ -1,5 +1,5 @@
-#ifndef ncplane_MAN_STRUCTURE
-#define ncplane_MAN_STRUCTURE
+#ifndef NOTCURSES_MAN_STRUCTURE
+#define NOTCURSES_MAN_STRUCTURE
 
 #ifdef __cplusplus
 extern "C" {
@@ -22,7 +22,10 @@ void docstructure_toggle(struct ncplane* p, struct ncplane* b, struct docstructu
 // add the specified [sub]section to the document strucure. |line| refers to
 // the row on the display plane, *not* the line in the original content.
 int docstructure_add(struct docstructure* ds, const char* title, int line,
-                     docstruct_e level);
+                     docstruct_e level, unsigned y);
+
+// update the docstructure browser based off a move of the page plane from to |newy|.
+int docstructure_move(struct docstructure* ds, unsigned newy);
 
 #ifdef __cplusplus
 }
